Simplifying things a little, and starting to modernize the code. Switched from homebrewed prime generation to using primesieve (requires installing and linking against primesieve).  Less code and slightly faster startup time.
Fixed long-standing bug in determination of number of sieve
workers -- ack.  Min and max were reversed.  That's just embarrassing.
Not clear this speeds things up yet, but it provides a basis for
actually tuning those values instead of having them always be 8.

Note to self:  Currently seeing 43.4 2ch/s on Haswell 4770 with
diff 1413.